Add more protections to 'Incomplete' Grades for Materialized Views ;

+++ db/CPG.sql
@@ -31,23 +31,23 @@ BEGIN
     FROM classes cls
     JOIN class_sections cs ON cls.class_id = cs.class_id
     WHERE cls.class_id = NEW.class_id;
-    AND NEW.grade != 'IN';
-
-    IF EXISTS (
-        SELECT 1 
-        FROM cpg_table 
-        WHERE course_id = retrieved_course_id
-        AND instructor = retrieved_instructor
-        AND grade = NEW.grade
-    ) THEN
-        UPDATE cpg_table 
-        SET grade_count = grade_count + 1
-        WHERE course_id = retrieved_course_id 
-        AND instructor = retrieved_instructor
-        AND grade = NEW.grade;
-    ELSE
-        INSERT INTO cpg_table (course_id, instructor, grade, grade_count)
-        VALUES (retrieved_course_id, retrieved_instructor, NEW.grade, 1);
+    IF NEW.grade != 'IN' THEN
+        IF EXISTS (
+            SELECT 1 
+            FROM cpg_table 
+            WHERE course_id = retrieved_course_id
+            AND instructor = retrieved_instructor
+            AND grade = NEW.grade
+        ) THEN
+            UPDATE cpg_table 
+            SET grade_count = grade_count + 1
+            WHERE course_id = retrieved_course_id 
+            AND instructor = retrieved_instructor
+            AND grade = NEW.grade;
+        ELSE
+            INSERT INTO cpg_table (course_id, instructor, grade, grade_count)
+            VALUES (retrieved_course_id, retrieved_instructor, NEW.grade, 1);
+        END IF;
     END IF;
     RETURN NEW;
 END;
@@ -70,25 +70,26 @@ BEGIN
     FROM classes cls
     JOIN class_sections cs ON cls.class_id = cs.class_id
     WHERE cls.class_id = OLD.class_id;
-
-    IF EXISTS (
-        SELECT 1 
-        FROM cpg_table 
-        WHERE course_id = retrieved_course_id 
-        AND instructor = retrieved_instructor 
-        AND grade = OLD.grade
-        AND grade_count > 1
-    ) THEN
-        UPDATE cpg_table
-        SET grade_count = grade_count - 1
-        WHERE course_id = retrieved_course_id 
-        AND instructor = retrieved_instructor 
-        AND grade = OLD.grade;
-    ELSE
-        DELETE FROM cpg_table
-        WHERE course_id = retrieved_course_id 
-        AND instructor = retrieved_instructor 
-        AND grade = OLD.grade;
+    IF OLD.grade != 'IN' THEN
+        IF EXISTS (
+            SELECT 1 
+            FROM cpg_table 
+            WHERE course_id = retrieved_course_id 
+            AND instructor = retrieved_instructor 
+            AND grade = OLD.grade
+            AND grade_count > 1
+        ) THEN
+            UPDATE cpg_table
+            SET grade_count = grade_count - 1
+            WHERE course_id = retrieved_course_id 
+            AND instructor = retrieved_instructor 
+            AND grade = OLD.grade;
+        ELSE
+            DELETE FROM cpg_table
+            WHERE course_id = retrieved_course_id 
+            AND instructor = retrieved_instructor 
+            AND grade = OLD.grade;
+        END IF;
     END IF;
     RETURN OLD;
 END;
